requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-bootstrap-v5==1.0.11
django-ckeditor==6.6.1
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
idna==3.4
mysql==0.0.3
mysqlclient==2.2.0
phonenumbers==8.13.19
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==6.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
